refactor: using ui-components package
