--- src/components/CollapseCard/types.ts
@@ -1,4 +1,0 @@
-export interface CollapseCardProps {
-  title: string;
-  body: string;
-}
