--- src/components/HomePage/HomePage.css
@@ -1,3 +1,0 @@
-.homeDesc {
-  text-align: left;
-}
