--- src/components/MapView/types.ts
@@ -1,3 +1,0 @@
-export interface MapViewProps {
-  searchText: string; // Search text for the map.
-}
